% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_theme.R
\name{ers_theme}
\alias{ers_theme}
\title{ERS Theme}
\source{
coppied from https://github.com/USDA-REE-ERS/MTED-Theme on 08/01/2025
}
\usage{
ers_theme()
}
\description{
ERS Theme
}
\examples{
ggplot2::ggplot() + ers_theme()
}
