% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makola-package.R
\docType{package}
\name{makola-package}
\alias{makola}
\alias{makola-package}
\title{makola: Tools and Workflows for Market Integration Analysis}
\description{
makola provides code, workflows, and supporting materials for market integration and related analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ftsiboe/makola}
  \item Report bugs at \url{https://github.com/ftsiboe/makola/issues}
}

}
\keyword{internal}
