% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE_GLOBALVARIABLES.R
\docType{data}
\name{PACKAGE_GLOBALVARIABLES}
\alias{PACKAGE_GLOBALVARIABLES}
\title{global_names}
\format{
A list of global names.
}
\source{
Internal innovation
}
\usage{
PACKAGE_GLOBALVARIABLES
}
\description{
A combined dataset for global_names
}
\keyword{datasets}
